# docs/swagger.yaml
definitions:
  ds.AuthResp:
    properties:
      token:
        type: string
    type: object
...
  ds.ErrorResp:
    properties:
      description:
        type: string
      status:
        type: string
    type: object
...
  ds.OkResp:
    properties:
      message:
        type: string
    type: object
...
  ds.UserRegisterReq:
    properties:
      login:
        type: string
      password:
        type: string
      repeat_password:
        type: string
    type: object
  ds.UserRegisterResp:
    properties:
      id:
        type: integer
      is_moderator:
        type: boolean
      login:
        type: string
    type: object
  ds.UserRespReq:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  ds.UserUpdateReq:
    properties:
      current_password:
        type: string
      new_password:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /user/login:
    post:
      consumes:
      - application/json
      description: Аутентификация пользователя и получение JWT-токена.
      parameters:
      - description: Данные для входа
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/ds.UserRespReq'
      produces:
      - application/json
      responses:
        "200":
          description: Токен JWT
          schema:
            $ref: '#/definitions/ds.AuthResp'
        "400":
          description: Неверный формат данных
          schema:
            $ref: '#/definitions/ds.ErrorResp'
        "401":
          description: Неверные учетные данные
          schema:
            $ref: '#/definitions/ds.ErrorResp'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/ds.ErrorResp'
      summary: Вход пользователя
      tags:
      - Auth
  /user/logout:
    post:
      description: Завершение текущей сессии пользователя.
      produces:
      - application/json
      responses:
        "200":
          description: Сообщение о выходе
          schema:
            $ref: '#/definitions/ds.OkResp'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/ds.ErrorResp'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/ds.ErrorResp'
      security:
      - ApiKeyAuth: []
      summary: Выход из системы
      tags:
      - Auth
  /user/reg:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя с указанными логином и паролем.
      parameters:
      - description: Данные для регистрации пользователя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/ds.UserRegisterReq'
      produces:
      - application/json
      responses:
        "201":
          description: Сообщение об успешной регистрации
          schema:
            $ref: '#/definitions/ds.OkResp'
        "400":
          description: Неверные данные или пароли не совпадают
          schema:
            $ref: '#/definitions/ds.ErrorResp'
        "409":
          description: Пользователь уже существует
          schema:
            $ref: '#/definitions/ds.ErrorResp'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/ds.ErrorResp'
      summary: Регистрация пользователя
      tags:
      - Auth
  /user/update:
    put:
      consumes:
      - application/json
      description: Обновляет данные текущего пользователя.
      parameters:
      - description: Новые данные пользователя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/ds.UserUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: Обновлённый пользователь
          schema:
            $ref: '#/definitions/ds.UserRegisterResp'
        "400":
          description: Некорректный формат данных
          schema:
            $ref: '#/definitions/ds.ErrorResp'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/ds.ErrorResp'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/ds.ErrorResp'
      security:
      - BearerAuth: []
      summary: Обновить данные пользователя
      tags:
      - Auth
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"